refactor: Moving around to avoid circular import
