+++ cognite/powerops/resync/models/base/asset_type.py
@@ -10,7 +10,7 @@ from cognite.client.data_classes import Relationship, TimeSeries, Label, Asset
 from pydantic import BaseModel
 
 from cognite.powerops.cdf_labels import AssetLabel, RelationshipLabel
-from .model import ResourceType
+from .resource_type import ResourceType
 from cognite.powerops.resync.models.cdf_resources import CDFSequence, CDFFile
 from cognite.powerops.resync.utils.serializer import get_pydantic_annotation
