+++ cognite/powerops/resync/models/base/model.py
@@ -2,70 +2,27 @@ from __future__ import annotations
 
 import abc
 from abc import ABC
-from typing import ClassVar, Callable, Iterable, Type as TypingType, Any, TypeVar, Literal, Union
+from typing import ClassVar, Callable, Iterable, Type as TypingType, Any, TypeVar, Literal
 
 from cognite.client import CogniteClient
 from cognite.client.data_classes._base import CogniteResource, CogniteResourceList
 from cognite.client.data_classes import (
     TimeSeriesList,
-    Asset,
-    Sequence,
-    FileMetadata,
     SequenceUpdate,
     FileMetadataUpdate,
 )
-from typing_extensions import Self, TypeAlias
+from typing_extensions import Self
 
 from cognite.client.data_classes import TimeSeries
 from pydantic import BaseModel
 from cognite.powerops.clients.powerops_client import PowerOpsClient
 from cognite.powerops.resync._changes import Change, FieldDifference, ModelDifference
+from cognite.powerops.resync.models.base.resource_type import _T_Type, ResourceType
 from cognite.powerops.resync.models.cdf_resources import CDFSequence, CDFFile, CDFResource
 from cognite.powerops.resync.models.helpers import isinstance_list
 from cognite.powerops.resync.utils.serializer import remove_read_only_fields
 from cognite.powerops.clients.data_classes._core import DomainModelApply
 from cognite.powerops.cogshop1.data_classes._core import DomainModelApply as DomainModelApplyCogShop1
-
-_T_Type = TypeVar("_T_Type")
-
-
-class ResourceType(BaseModel, ABC):
-    def sequences(self) -> list[CDFSequence]:
-        return self._fields_of_type(CDFSequence)
-
-    def files(self) -> list[CDFFile]:
-        return self._fields_of_type(CDFFile)
-
-    def time_series(self) -> list[TimeSeries]:
-        return self._fields_of_type(TimeSeries)
-
-    def _fields_of_type(self, type_: TypingType[_T_Type]) -> list[_T_Type]:
-        output: list[_T_Type] = []
-        for field_name in self.model_fields:
-            value = getattr(self, field_name)
-            if not value:
-                continue
-            elif isinstance_list(value, type_):
-                output.extend(value)
-            elif isinstance(value, type_):
-                output.append(value)
-        return output
-
-    @property
-    def external_id(self) -> str:
-        raise NotImplementedError()
-
-    def standardize(self) -> None:
-        """
-        This ensures that the model is in a standardized form.
-
-        This is useful when doing comparisons between models, as for example, the ordering of the assets in some
-        lists does not matter.
-        """
-        ...
-
-
-Resource: TypeAlias = Union[Asset, TimeSeries, Sequence, FileMetadata, ResourceType]
 
 
 class Model(BaseModel, ABC):
